% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Visualization_Help.R
\name{Clust_Alg_Plot}
\alias{Clust_Alg_Plot}
\title{CLUSTERING ALGORITHM VISUALIZATION FUNCTION}
\usage{
Clust_Alg_Plot(Entry, Option)
}
\arguments{
\item{Entry}{List containing rankings and dictionary of the algoritm's configurations previously calculated with the function Clust_Alg_Rank}

\item{Option}{Object to visualize}
}
\value{
Visualizations of mean ranking, most times winner ranking, best mean algorithm, most times winner algorithm and dictionary with the algorithms and configurations
}
\description{
Visualize clustering rankings and winners from the different algoritm's configurations
}
\examples{
Option <- "Mean_Ranking"
Clust_Alg_Plot(Entry, Option)

Option <- "Count_Best_Ranking"
Clust_Alg_Plot(Entry, Option)

Option <- "Mean_Best_Algorithm"
Clust_Alg_Plot(Entry, Option)

Option <- "Count_Best_Algorithm"
Clust_Alg_Plot(Entry, Option)

Option <- "Dictionary"
Clust_Alg_Plot(Entry, Option)
}
