% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_Algorithms_Ranking_Help.R
\name{Clust_Alg_Rank}
\alias{Clust_Alg_Rank}
\title{CLUSTERING ALGORITHM RANKING FUNCTION}
\usage{
Clust_Alg_Rank(Dataset, Algorithm_Configs, Metrics_List)
}
\arguments{
\item{Dataset}{Normalized matrix of continous data}

\item{Algorithm_Configs}{List of algorithms and configurations}

\item{Metrics_List}{List of metrics for evaluation}
}
\value{
List with mean ranking, most times winner ranking and dictionary with the algorithms and configurations
}
\description{
Calculate clustering rankings for the different algoritm's configurations
}
\examples{
Dataset <- USArrests
Dataset <- na.omit(Dataset)
Dataset <- scale(Dataset)
Algorithm_Configs <- list(km = list(name = "kmeans", config = list(centers = c(3,4,5), iter.max = 10)), cm = list(name = "Cluster_Medoids", config = list(clusters = c(4,5,6), swap_phase = TRUE)))
Metrics_List <- "all"
Entry <- Clust_Alg_Rank(Dataset, Algorithm_Configs, Metrics_List)
}
