% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_elements.R
\name{knn_elements}
\alias{knn_elements}
\title{'Elements' matrix computation}
\usage{
knn_elements(y, d)
}
\arguments{
\item{y}{A matrix.}

\item{d}{Length of each of the 'elements'.}
}
\value{
A matrix to be used for calculating distances.
}
\description{
Creates a matrix to be used for calculating distances. The most
recent 'element' is put in the first row of the matrix, the
second most recent 'element' in the second row and so on. Therefore,
the oldest 'element' is put in the last row.
}
